# BCM/eAarogya-Portal-for-EHR-Management-master/eAarogya-Portal-for-EHR-Management-master/ehr-api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Alpha",
        "version": "1.0.0",
        "title": "eMedical Record System"
    },
    "host": "localhost:3000",
    "basePath": "/ehr",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/addReport": {
            "post": {
                "summary": "Add a new medical report",
                "description": "Add Report",
                "operationId": "addReport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Report object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/addReport"
                    }
                }],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/addMedicineReport": {
            "post": {
                "summary": "Add a new medicine prescription report",
                "description": "Add Prescription",
                "operationId": "addMedicineReport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Report object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/addMedicineReport"
                    }
                }],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/addrLReport": {
            "post": {
                "summary": "Add a new RadioLogist report",
                "description": "Add Report and links",
                "operationId": "addrLReport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Report object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/addrLReport"
                    }
                }],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/getReport": {
            "post": {
                "summary": "Get the details of the previous report",
                "description": "Get report",
                "operationId": "getReport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Report object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/getReport"
                    }
                }],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/getRecord": {
            "post": {
                "summary": "Get the history of the record",
                "description": "Get record",
                "operationId": "getRecord",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Record object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/getReport"
                    }
                }],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/getMedicineReport": {
            "post": {
                "summary": "Get the details of the latest prescription",
                "description": "Get prescription report",
                "operationId": "getMedicineReport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Report object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/getReport"
                    }
                }],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/getMedicineRecord": {
            "post": {
                "summary": "Get the details of the history of prescriptions",
                "description": "Get prescription history",
                "operationId": "getMedicineRecord",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Report object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/getReport"
                    }
                }],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "addReport": {
            "type": "object",
            "properties": {
                "recordID": {
                    "type": "string",
                    "description": "Existing Record ID"
                },
                "report": {
                    "type": "string",
                    "description": "Medical report"
                }
            }
        },
        "addMedicineReport": {
            "type": "object",
            "properties": {
                "recordID": {
                    "type": "string",
                    "description": "Existing Record ID"
                },
                "prescription": {
                    "type": "string",
                    "description": "Prescription"
                }
            }
        },
        "addrLReport": {
            "type": "object",
            "properties": {
                "recordID": {
                    "type": "string",
                    "description": "Existing Record ID"
                },
                "report": {
                    "type": "string",
                    "description": "Medical report"
                },
                "links": {
                    "type": "string",
                    "description": "Links to report images"
                }
            }
        },
        "getReport": {
            "type": "object",
            "properties": {
                "recordID": {
                    "type": "string",
                    "description": "Exising record ID"
                }
            }
        }
    },
...
}